DB版に着手して気になった点を修正した
- erbへ変数を渡す方法をテンプレート変数に変更
- Notebookクラスを別ファイルにした
