+++ views/detail.erb
@@ -1,17 +1,17 @@
 <section class="p-5 bg-white border rounded-3 mb-4">
   <div class="row">
     <div class="col-10">
-      <h2><%= note[:title] %></h2>
+      <h2><%= @note[:title] %></h2>
     </div>
     <div class="col-2 text-end">
-      <form action="/notes/<%= note[:id] %>" method="post">
+      <form action="/notes/<%= @note[:id] %>" method="post">
         <input type="hidden" name="_method" value="DELETE">
         <button type="submit" class="btn btn-lg btn-outline-danger me-3"><i class="bi bi-trash3"></i></button>
-        <a class="btn btn-lg btn-outline-dark" href="/notes/<%= note[:id] %>/edit" role="button"><i class="bi bi-pen-fill"></i> Edit</a>
+        <a class="btn btn-lg btn-outline-dark" href="/notes/<%= @note[:id] %>/edit" role="button"><i class="bi bi-pen-fill"></i> Edit</a>
       </form>
     </div>
   </div>
   <hr>
-  <div class="fs-5"><%= note[:text].gsub(/(\r\n|\r|\n)/, "<br>") %></div>
+  <div class="fs-5"><%= @note[:text].gsub(/(\r\n|\r|\n)/, "<br>") %></div>
 </section>
 <a class="btn btn-lg btn-outline-dark" href="/notes">Back</a>
