+++ views/edit.erb
@@ -1,10 +1,10 @@
-<form method="post" action="/notes/<%= note[:id] %>">
+<form method="post" action="/notes/<%= @note[:id] %>">
   <input type="hidden" name="_method" value="patch">
-  <input class="form-control form-control-lg mb-4" type="text" name="title" value="<%= note[:title]%>">
-  <textarea class="form-control form-control-lg mb-4" name="text" rows="20"><%= note[:text] %></textarea>
+  <input class="form-control form-control-lg mb-4" type="text" name="title" value="<%= @note[:title]%>">
+  <textarea class="form-control form-control-lg mb-4" name="text" rows="20"><%= @note[:text] %></textarea>
   <div class="row">
     <div class="col">
-      <a class="btn btn-lg btn-outline-dark" href="/notes/<%= note[:id] %>">Back</a>
+      <a class="btn btn-lg btn-outline-dark" href="/notes/<%= @note[:id] %>">Back</a>
     </div>
     <div class="col text-end mb-4">
       <button class="btn btn-lg btn-dark" type="submit">Save</button>
